Fix file names with slashes (#450)
In Luigi's Manson 3, some of the archived models have slashes in their names. This means that they can't be exported with those names, and Toolbox will crash when it reaches them, if you try to "export all".

What i have done to fix this, is make toolbox treat names with slashes as directory paths, with a file name on the end.