ADD - LZ77(Wii type 11) Compression
This pull request adds the ability to compress files to the .lz format of type 11 for Wii.
This functionality has been enabled in the Tools > Compressions > LZ77(Wii type 11) > Compress tab.
It has been successfully tested with the game Paper Mario: Color Splash for the Wii U.
It is important to note that this is a modified version of dsdecmp.

Also changed was the SaveFileForCompression method to remove the file extension when decompressing files with multiple extensions. Previously, the method would add a ".dec" extension to the file name when decompressing it. With this change, the method now removes the original file extension when decompressing a file with multiple extensions. These changes should improve the usability of the SaveFileForCompression method by ensuring that decompressed files with multiple extensions are saved with the correct file name.