Export material data as .json
Works with:
- Right click material -> Export
- Right click "Materials" folder -> "Export All Materials"
- Tools -> Batch Export Models

Currently it's not optional, and the amount of data is pretty significant,
but I'll leave that to you if you want to make it optional.

I simplified some structures to make the data more usable, but
I did not try to filter anything so as to avoid incorrect assumptions
about what may or may not be useful.